REF: split data into data_base and data_active/data_passive
