+++ atef/widgets/config/data_passive.py
@@ -1,474 +1,39 @@
 """
-Widgets used for manipulating the configuration data.
+Widgets used for manipulating Passive Checkout Data
 """
-from __future__ import annotations
 
 from collections import defaultdict
 from dataclasses import dataclass
-from typing import (Any, Callable, ClassVar, Dict, List, Optional, Protocol,
-                    Tuple)
-from weakref import WeakValueDictionary
+from typing import Any, Callable, Dict, List, Optional
 
 from pydm.widgets.drawing import PyDMDrawingLine
 from qtpy.QtGui import QColor, QDropEvent
-from qtpy.QtWidgets import (QCheckBox, QComboBox, QFrame, QHBoxLayout, QLabel,
-                            QLayout, QLineEdit, QMessageBox, QPlainTextEdit,
-                            QPushButton, QSpinBox, QStyle, QTableWidget,
-                            QTableWidgetItem, QToolButton, QVBoxLayout,
-                            QWidget)
+from qtpy.QtWidgets import (QCheckBox, QComboBox, QFrame, QLabel, QLineEdit,
+                            QMessageBox, QPushButton, QSpinBox, QStyle,
+                            QTableWidget, QTableWidgetItem, QToolButton,
+                            QVBoxLayout)
 
 from atef.check import Comparison, Equals, Value
 from atef.config import (Configuration, ConfigurationGroup,
-                         DeviceConfiguration, GroupResultMode, PVConfiguration,
-                         ToolConfiguration)
+                         DeviceConfiguration, GroupResultMode, PVConfiguration)
 from atef.enums import Severity
-from atef.qt_helpers import QDataclassBridge, QDataclassList
+from atef.qt_helpers import QDataclassList
 from atef.reduce import ReduceMethod
 from atef.tools import Ping
 from atef.type_hints import PrimitiveType
-from atef.widgets.archive_viewer import get_archive_viewer
+from atef.widgets.config.data_base import DataWidget, SimpleRowWidget
+from atef.widgets.config.utils import (BulkListWidget, ComponentListWidget,
+                                       DeviceListWidget, setup_line_edit_data,
+                                       user_string_to_bool)
 from atef.widgets.core import DesignerDisplay
-from atef.widgets.utils import FrameOnEditFilter, match_line_edit_text_width
-
-from .utils import (BulkListWidget, ComponentListWidget, DeviceListWidget,
-                    get_relevant_pvs, setup_line_edit_data,
-                    user_string_to_bool)
-
-
-class AnyDataclass(Protocol):
-    """
-    Protocol stub shamelessly lifted from stackoverflow to hint at dataclass
-    """
-    __dataclass_fields__: Dict
-
-
-class DataWidget(QWidget):
-    """
-    Base class for widgets that manipulate dataclasses.
-
-    Defines the init args for all data widgets and handles synchronization
-    of the ``QDataclassBridge`` instances. This is done so that only data
-    widgets need to consider how to handle bridges and the page classes
-    simply need to pass in data structures, rather than needing to keep track
-    of how two widgets editing the same data structure must share the same
-    bridge object.
-
-    Parameters
-    ----------
-    data : any dataclass
-        The dataclass that the widget needs to manipulate. Most widgets are
-        expecting either specific dataclasses or dataclasses that have
-        specific matching fields.
-    kwargs : QWidget kwargs
-        Passed directly to QWidget's __init__. Likely unused in most cases.
-        Even parent is unlikely to see use because parent is set automatically
-        when a widget is inserted into a layout.
-    """
-    _bridge_cache: ClassVar[
-        WeakValueDictionary[int, QDataclassBridge]
-    ] = WeakValueDictionary()
-    bridge: QDataclassBridge
-    data: AnyDataclass
-
-    def __init__(self, data: AnyDataclass, **kwargs):
-        super().__init__(**kwargs)
-        self.data = data
-        try:
-            # TODO figure out better way to cache these
-            # TODO worried about strange deallocation timing race conditions
-            self.bridge = self._bridge_cache[id(data)]
-        except KeyError:
-            bridge = QDataclassBridge(data)
-            self._bridge_cache[id(data)] = bridge
-            self.bridge = bridge
-
-
-class NameMixin:
-    """
-    Mixin class for distributing init_name
-    """
-    def init_name(self) -> None:
-        """
-        Set up the name_edit widget appropriately.
-        """
-        # Load starting text
-        load_name = self.bridge.name.get() or ''
-        self.last_name = load_name
-        self.name_edit.setText(load_name)
-        # Set up the saving/loading
-        self.name_edit.textEdited.connect(self.update_saved_name)
-        self.bridge.name.changed_value.connect(self.apply_new_name)
-
-    def update_saved_name(self, name: str) -> None:
-        """
-        When the user edits the name, write to the config.
-        """
-        self.last_name = self.name_edit.text()
-        self.bridge.name.put(name)
-
-    def apply_new_name(self, text: str) -> None:
-        """
-        If the text changed in the data, update the widget.
-
-        Only run if needed to avoid annoyance with cursor repositioning.
-        """
-        if text != self.last_name:
-            self.name_edit.setText(text)
-
-
-class NameDescTagsWidget(DesignerDisplay, NameMixin, DataWidget):
-    """
-    Widget for displaying and editing the name, description, and tags fields.
-
-    Any of these will be automatically disabled if the data source is missing
-    the corresponding field.
-
-    As a convenience, this widget also holds a parent_button in a convenient
-    place for page layouts, since it is expected that this will be near the
-    top of the page, and an "extra_text_label" QLabel for general use.
-    """
-    filename = 'name_desc_tags_widget.ui'
-
-    name_edit: QLineEdit
-    name_frame: QFrame
-    desc_edit: QPlainTextEdit
-    desc_frame: QFrame
-    tags_content: QVBoxLayout
-    add_tag_button: QToolButton
-    tags_frame: QFrame
-    parent_button: QToolButton
-    action_button: QToolButton
-    extra_text_label: QLabel
-
-    last_name: str
-    last_desc: str
-    pvs: List[Tuple[str, str]]  # (pv, attrname)
-
-    def __init__(self, data: AnyDataclass, **kwargs):
-        super().__init__(data=data, **kwargs)
-        try:
-            self.bridge.name
-        except AttributeError:
-            self.name_frame.hide()
-        else:
-            self.init_name()
-        try:
-            self.bridge.description
-        except AttributeError:
-            self.desc_frame.hide()
-        else:
-            self.init_desc()
-        try:
-            self.bridge.tags
-        except AttributeError:
-            self.tags_frame.hide()
-        else:
-            self.init_tags()
-
-        # if there's a pv, show the button for archive widget.
-        # info would be filled in after init... don't show at start
-        self.action_button.hide()
-        self._viewer_initialized = False
-
-    def init_desc(self) -> None:
-        """
-        Set up the desc_edit widget appropriately.
-        """
-        # Load starting text
-        load_desc = self.bridge.description.get() or ''
-        self.last_desc = load_desc
-        self.desc_edit.setPlainText(load_desc)
-        # Setup the saving/loading
-        self.desc_edit.textChanged.connect(self.update_saved_desc)
-        self.bridge.description.changed_value.connect(self.apply_new_desc)
-        self.desc_edit.textChanged.connect(self.update_text_height)
-
-    def update_saved_desc(self) -> None:
-        """
-        When the user edits the desc, write to the config.
-        """
-        self.last_desc = self.desc_edit.toPlainText()
-        self.bridge.description.put(self.last_desc)
-
-    def apply_new_desc(self, desc: str) -> None:
-        """
-        When some other widget updates the description, update it here.
-        """
-        if desc != self.last_desc:
-            self.desc_edit.setPlainText(desc)
-
-    def showEvent(self, *args, **kwargs) -> None:
-        """
-        Override showEvent to update the desc height when we are shown.
-        """
-        try:
-            self.update_text_height()
-        except AttributeError:
-            pass
-        return super().showEvent(*args, **kwargs)
-
-    def resizeEvent(self, *args, **kwargs) -> None:
-        """
-        Override resizeEvent to update the desc height when we resize.
-        """
-        try:
-            self.update_text_height()
-        except AttributeError:
-            pass
-        return super().resizeEvent(*args, **kwargs)
-
-    def update_text_height(self) -> None:
-        """
-        When the user edits the desc, make the text box the correct height.
-        """
-        line_count = max(self.desc_edit.document().size().toSize().height(), 1)
-        self.desc_edit.setFixedHeight(line_count * 13 + 12)
-
-    def init_tags(self) -> None:
-        """
-        Set up the various tags widgets appropriately.
-        """
-        tags_list = TagsWidget(
-            data_list=self.bridge.tags,
-            layout=QHBoxLayout(),
-        )
-        self.tags_content.addWidget(tags_list)
-
-        def add_tag() -> None:
-            if tags_list.widgets and not tags_list.widgets[-1].line_edit.text().strip():
-                # Don't add another tag if we haven't filled out the last one
-                return
-
-            elem = tags_list.add_item('')
-            elem.line_edit.setFocus()
-
-        self.add_tag_button.clicked.connect(add_tag)
-
-    def init_viewer(self, attr: str, config: Configuration) -> None:
-        """ Set up the archive viewer button """
-        if self._viewer_initialized:
-            # make sure this only happens once per instance
-            return
-
-        if ((hasattr(config, 'by_attr') or hasattr(config, 'by_pv'))
-                and not isinstance(config, ToolConfiguration)):
-            icon = self.style().standardIcon(QStyle.SP_FileDialogContentsView)
-            self.action_button.setIcon(icon)
-            self.action_button.setToolTip('Open Archive Viewer with '
-                                          'relevant signals')
-            self.action_button.show()
-
-        def open_arch_viewer(*args, **kwargs):
-            # only query PV info once requested.  grabbing devices and
-            # their relevant PVs can be time consuming
-            pv_list = get_relevant_pvs(attr, config)
-            if len(pv_list) == 0:
-                QMessageBox.information(
-                    self,
-                    'No Archived PVs',
-                    'No valid PVs found to plot with archive viewer. '
-                    'Signal may be a derived signal'
-                )
-                self.action_button.hide()
-                return
-            widget = get_archive_viewer()
-            for pv, dev_attr in pv_list:
-                widget.add_signal(pv, dev_attr=dev_attr, update_curves=False)
-            widget.update_curves()
-            widget.show()
-
-        self.action_button.clicked.connect(open_arch_viewer)
-        self._viewer_initialized = True
-
-
-class TagsWidget(QWidget):
-    """
-    A widget used to edit a QDataclassList tags field.
-
-    Aims to emulate the look and feel of typical tags fields
-    in online applications.
-
-    Parameters
-    ----------
-    data_list : QDataclassList
-        The dataclass list to edit using this widget.
-    layout : QLayout
-        The layout to use to arrange our labels. This should be an
-        instantiated but not placed layout. This lets us have some
-        flexibility in whether we arrange things horizontally,
-        vertically, etc.
-    """
-    widgets: List[TagsElem]
-
-    def __init__(
-        self,
-        data_list: QDataclassList,
-        layout: QLayout,
-        *args,
-        **kwargs,
-    ):
-        super().__init__(*args, **kwargs)
-        self.data_list = data_list
-        self.setLayout(layout)
-        self.widgets = []
-        starting_list = data_list.get()
-        if starting_list is not None:
-            for starting_value in starting_list:
-                self.add_item(starting_value, init=True)
-
-    def add_item(
-        self,
-        starting_value: str,
-        init: bool = False,
-        **kwargs,
-    ) -> TagsElem:
-        """
-        Create and add new editable widget element to this widget's layout.
-
-        This can either be an existing string on the dataclass list to keep
-        track of, or it can be used to add a new string to the dataclass list.
-
-        This method will also set up the signals and slots for the new widget.
-
-        Parameters
-        ----------
-        starting_value : str
-            The starting text value for the new widget element.
-            This should match the text exactly for tracking existing
-            strings.
-        checked : bool, optional
-            This argument is unused, but it will be sent by various button
-            widgets via the "clicked" signal so it must be present.
-        init : bool, optional
-            Whether or not this is the initial initialization of this widget.
-            This will be set to True in __init__ so that we don't mutate
-            the underlying dataclass. False, the default, means that we're
-            adding a new string to the dataclass, which means we should
-            definitely append it.
-        **kwargs : from qt signals
-            Other kwargs sent along with qt signals will be ignored.
-
-        Returns
-        -------
-        strlistelem : StrListElem
-            The widget created by this function call.
-        """
-        new_widget = TagsElem(starting_value, self)
-        self.widgets.append(new_widget)
-        if not init:
-            self.data_list.append(starting_value)
-        self.layout().addWidget(new_widget)
-        return new_widget
-
-    def save_item_update(self, item: TagsElem, new_value: str) -> None:
-        """
-        Update the dataclass as appropriate when the user submits a new value.
-
-        Parameters
-        ----------
-        item : StrListElem
-            The widget that the user has edited.
-        new_value : str
-            The value that the user has submitted.
-        """
-        index = self.widgets.index(item)
-        self.data_list.put_to_index(index, new_value)
-
-    def remove_item(self, item: TagsElem) -> None:
-        """
-        Update the dataclass as appropriate when the user removes a value.
-
-        Parameters
-        ----------
-        item : StrListElem
-            The widget that the user has clicked the delete button for.
-        """
-        index = self.widgets.index(item)
-        self.widgets.remove(item)
-        self.data_list.remove_index(index)
-        item.deleteLater()
-
-
-class TagsElem(DesignerDisplay, QWidget):
-    """
-    A single element for the TagsWidget.
-
-    Has a QLineEdit for changing the text and a delete button.
-    Changes its style to no frame when it has text and is out of focus.
-    Only shows the delete button when the text is empty.
-
-    Parameters
-    ----------
-    start_text : str
-        The starting text for this tag.
-    tags_widget : TagsWidget
-        A reference to the TagsWidget that contains this widget.
-    """
-    filename = 'tags_elem.ui'
-
-    line_edit: QLineEdit
-    del_button: QToolButton
-
-    def __init__(self, start_text: str, tags_widget: TagsWidget, **kwargs):
-        super().__init__(**kwargs)
-        self.line_edit.setText(start_text)
-        self.tags_widget = tags_widget
-        edit_filter = FrameOnEditFilter(parent=self)
-        edit_filter.set_no_edit_style(self.line_edit)
-        self.line_edit.installEventFilter(edit_filter)
-        self.on_text_changed(start_text)
-        self.line_edit.textChanged.connect(self.on_text_changed)
-        self.line_edit.textEdited.connect(self.on_text_edited)
-        self.del_button.clicked.connect(self.on_del_clicked)
-        icon = self.style().standardIcon(QStyle.SP_TitleBarCloseButton)
-        self.del_button.setIcon(icon)
-
-    def on_text_changed(self, text: str) -> None:
-        """
-        Edit our various visual elements when the text changes.
-
-        This will do all of the following:
-        - make the delete button show only when the text field is empty
-        - adjust the size of the text field to be roughly the size of the
-          string we've inputted
-        """
-        # Show or hide the del button as needed
-        self.del_button.setVisible(not text)
-        # Adjust the width to match the text
-        match_line_edit_text_width(self.line_edit, text=text)
-
-    def on_data_changed(self, data: str) -> None:
-        """
-        Change the text displayed here using new data, if needed.
-        """
-        if self.line_edit.text() != data:
-            self.line_edit.setText(data)
-
-    def on_text_edited(self, text: str) -> None:
-        """
-        Update the dataclass when the user edits the text.
-        """
-        self.tags_widget.save_item_update(
-            item=self,
-            new_value=text,
-        )
-
-    def on_del_clicked(self, **kwargs) -> None:
-        """
-        Tell the QTagsWidget when our delete button is clicked.
-        """
-        self.tags_widget.remove_item(self)
 
 
 class ConfigurationGroupWidget(DesignerDisplay, DataWidget):
     """
     Widget for modifying most unique fields in ConfigurationGroup.
-
     The fields handled here are:
-
     - values: dict[str, Any]
     - mode: GroupResultMode
-
     The configs field will be modified by the ConfigurationGroupRowWidget,
     which is intended to be used many times, once each to handle each
     sub-Configuration instance.
@@ -528,9 +93,7 @@ class ConfigurationGroupWidget(DesignerDisplay, DataWidget):
     ) -> None:
         """
         Adds a new value to the global values table.
-
         The default value is an empty string name with an empty string value.
-
         Parameters
         ----------
         name : str, optional
@@ -564,7 +127,6 @@ class ConfigurationGroupWidget(DesignerDisplay, DataWidget):
     def resize_table(self) -> None:
         """
         Set the table to a fixed height to show the available rows.
-
         This allows us to contract the table's size when there are few rows,
         increase the size as we add rows, and set a maximum size after which
         the table will acquire a scrollbar.
@@ -588,11 +150,9 @@ class ConfigurationGroupWidget(DesignerDisplay, DataWidget):
     def on_table_edit(self, row: int, column: int) -> None:
         """
         Slot for updating the saved values when the table is edited.
-
         Regardless of which row or column is changed, we'll reconstruct
         the entire values dictionary to make sure it is serialized with
         consistent ordering.
-
         The arguments are passed by the qt signal but are unused.
         """
         if self.adding_new_row:
@@ -655,20 +215,14 @@ class ConfigurationGroupWidget(DesignerDisplay, DataWidget):
 class DeviceConfigurationWidget(DesignerDisplay, DataWidget):
     """
     Handle the unique static fields from DeviceConfiguration.
-
     The fields handled fully here are:
-
     - devices: List[str]
-
     The fields handled partially here are:
-
     - by_attr: Dict[str, List[Comparison]]
     - shared: List[Comparison] = field(default_factory=list)
-
     This will only put empty lists into the by_attr dict.
     Filling those lists will be the responsibility of the
     DeviceConfigurationPageWidget.
-
     The shared list will be used a place to put configurations
     that have had their attr deleted instead of just dropping
     those entirely, but adding to the shared list will normally
@@ -712,10 +266,8 @@ class DeviceConfigurationWidget(DesignerDisplay, DataWidget):
     def add_new_signal(self, name: str) -> None:
         """
         Add a new signal for use in the comparison selectors.
-
         These are stored as the keys in the by_attr dictionary.
         A new signal should start as mapping to an empty list.
-
         Parameters
         ----------
         name : str
@@ -729,12 +281,10 @@ class DeviceConfigurationWidget(DesignerDisplay, DataWidget):
     def remove_signal(self, name: str) -> None:
         """
         Remove an existing signal from the usage pool for the comparisons.
-
         These are stored as the keys in the by_attr dictionary.
         When we remove a signal, any orphaned comparisons should
         migrate over to the "shared" comparisons list to avoid losing
         them.
-
         Parameters
         ----------
         name : str
@@ -759,7 +309,6 @@ class DeviceConfigurationWidget(DesignerDisplay, DataWidget):
 class ListHolder:
     """
     Dummy dataclass to match ComponentListWidget API
-
     Previous versions of the application used lists to store
     things like signals, etc., and the widgets written for this
     version assume that a dataclass with a list attribute exists.
@@ -770,16 +319,12 @@ class ListHolder:
 class PVConfigurationWidget(DataWidget):
     """
     Handle the unique static fields from PVConfiguration.
-
     The fields handled partially here are:
-
     - by_pv: Dict[str, List[Comparison]]
     - shared: List[Comparison] = field(default_factory=list)
-
     This will only put empty lists into the by_pv dict.
     Filling those lists will be the responsibility of the
     PVConfigurationPageWidget.
-
     The shared list will be used a place to put configurations
     that have had their pv deleted instead of just dropping
     those entirely, but adding to the shared list will normally
@@ -815,10 +360,8 @@ class PVConfigurationWidget(DataWidget):
     def add_new_signal(self, name: str) -> None:
         """
         Add a new pv for use in the comparison selectors.
-
         These are stored as the keys in the by_pv dictionary.
         A new signal should start as mapping to an empty list.
-
         Parameters
         ----------
         name : str
@@ -832,12 +375,10 @@ class PVConfigurationWidget(DataWidget):
     def remove_signal(self, name: str) -> None:
         """
         Remove an existing pv from the usage pool for the comparisons.
-
         These are stored as the keys in the by_pv dictionary.
         When we remove a pv, any orphaned comparisons should
         migrate over to the "shared" comparisons list to avoid losing
         them.
-
         Parameters
         ----------
         name : str
@@ -861,12 +402,10 @@ class PVConfigurationWidget(DataWidget):
 class PingWidget(DesignerDisplay, DataWidget):
     """
     Widget that modifies the fields in the Ping tool.
-
     These fields are:
     - hosts: List[str] = field(default_factory=list)
     - count: int = 3
     - encoding: str = "utf-8"
-
     This will include a list widget on the left for the
     hosts and a basic form on the right for the other
     fields.
@@ -907,51 +446,13 @@ class PingWidget(DesignerDisplay, DataWidget):
         self.bridge.count.put(self.count_spinbox.value())
 
 
-class SimpleRowWidget(NameMixin, DataWidget):
-    """
-    Common behavior for these simple rows included on the various pages.
-    """
-    name_edit: QLineEdit
-    child_button: QToolButton
-    delete_button: QToolButton
-
-    def setup_row(self) -> None:
-        """
-        Make the commonalities in simple row widgets functional.
-        """
-        self.init_name()
-        self.edit_filter = FrameOnEditFilter(parent=self)
-        self.name_edit.installEventFilter(self.edit_filter)
-        self.name_edit.textChanged.connect(self.on_name_edit_text_changed)
-        self.on_name_edit_text_changed()
-
-    def adjust_edit_filter(self) -> None:
-        """
-        Toggle between edit/no edit style modes based on having a valid name.
-        """
-        if self.bridge.name.get():
-            self.edit_filter.set_no_edit_style(self.name_edit)
-        else:
-            self.edit_filter.set_edit_style(self.name_edit)
-
-    def on_name_edit_text_changed(self, **kwargs) -> None:
-        """
-        Updates the style of our name edit appropriately on text change.
-        """
-        match_line_edit_text_width(self.name_edit)
-        if not self.name_edit.hasFocus():
-            self.adjust_edit_filter()
-
-
 class ConfigurationGroupRowWidget(DesignerDisplay, SimpleRowWidget):
     """
     A row summary of a ``Configuration`` instance of a ``ConfigurationGroup``.
-
     You can view and edit the name from here, or delete the row.
     This will also show the class of the configuration, e.g. if it
     is a DeviceConfiguration for example, and will provide a
     button for navigation to the correct child page.
-
     The child_button and delete_button need to be set up by the page that
     includes this widget, as this widget has no knowledge of page navigation
     or of data outside of its ``Configuration`` instance, so it can't
@@ -971,7 +472,6 @@ class ConfigurationGroupRowWidget(DesignerDisplay, SimpleRowWidget):
 class ComparisonRowWidget(DesignerDisplay, SimpleRowWidget):
     """
     Handle one comparison instance embedded on a configuration page.
-
     The attr_combo is controlled by the page this is placed in.
     It may be a PV, it may be a signal, it may be a ping result, and
     it might be a key value like "shared" with special meaning.
@@ -1066,9 +566,7 @@ class GeneralComparisonWidget(DesignerDisplay, DataWidget):
     def new_invert_combo(self, index: int) -> None:
         """
         Slot to handle user input in the generic "Invert" combo box.
-
         Uses the current bridge to mutate the stored dataclass.
-
         Parameters
         ----------
         index : int
@@ -1079,12 +577,9 @@ class GeneralComparisonWidget(DesignerDisplay, DataWidget):
     def new_reduce_period_edit(self, value: str) -> None:
         """
         Slot to handle user intput in the generic "Reduce Period" line edit.
-
         Tries to interpet user input as a float. If this is not possible,
         the period will not be updated.
-
         Uses the current bridge to mutate the stored dataclass.
-
         Parameters
         ----------
         value : str
@@ -1100,9 +595,7 @@ class GeneralComparisonWidget(DesignerDisplay, DataWidget):
     def new_reduce_method_combo(self, value: str) -> None:
         """
         Slot to handle user input in the generic "Reduce Method" combo box.
-
         Uses the current bridge to mutate the stored dataclass.
-
         Parameters
         ----------
         value : str
@@ -1113,9 +606,7 @@ class GeneralComparisonWidget(DesignerDisplay, DataWidget):
     def new_string_combo(self, index: int) -> None:
         """
         Slot to handle user input in the generic "String" combo box.
-
         Uses the current bridge to mutate the stored dataclass.
-
         Parameters
         ----------
         index : int
@@ -1126,9 +617,7 @@ class GeneralComparisonWidget(DesignerDisplay, DataWidget):
     def new_sev_on_failure_combo(self, value: str) -> None:
         """
         Slot to handle user input in the "Severity on Failure" combo box.
-
         Uses the current bridge to mutate the stored dataclass.
-
         Parameters
         ----------
         value : str
@@ -1139,9 +628,7 @@ class GeneralComparisonWidget(DesignerDisplay, DataWidget):
     def new_if_disc_combo(self, value: str) -> None:
         """
         Slot to handle user input in the "If Disconnected" combo box.
-
         Uses the current bridge to mutate the stored dataclass.
-
         Parameters
         ----------
         value : str
@@ -1153,7 +640,6 @@ class GeneralComparisonWidget(DesignerDisplay, DataWidget):
 class EqualsMixin:
     """
     Utilities for atol/rtol style data widgets
-
     Used in EqualsWidget and ValueRowWidget
     """
     label_to_type: Dict[str, type] = {
@@ -1182,7 +668,6 @@ class EqualsMixin:
     def setup_equals_widget(self) -> None:
         """
         Do all the setup needed to make this widget functional.
-
         Things handled here:
         - Set up the data type selection to know whether or not
           atol/rtol/range means anything and so that we can allow
@@ -1226,11 +711,9 @@ class EqualsMixin:
     def update_range_label(self, *args, **kwargs) -> None:
         """
         Update the range label as appropriate.
-
         If our value is an int or float, this will do calculations
         using the atol and rtol to report the tolerance
         of the range to the user.
-
         If our value is a bool, this will summarize whether our
         value is being interpretted as True or False.
         """
@@ -1254,14 +737,12 @@ class EqualsMixin:
     ) -> PrimitiveType:
         """
         Convert our line edit value into a string based on the combobox.
-
         Parameters
         ----------
         value : str, optional
             The text contents of our line edit.
         gui_type_str : str, optional
             The text contents of our combobox.
-
         Returns
         -------
         converted : Any
@@ -1277,13 +758,10 @@ class EqualsMixin:
     def new_gui_type(self, gui_type_str: str) -> None:
         """
         Slot for when the user changes the GUI data type.
-
         Re-interprets our value as the selected type. This will
         update the current value in the bridge as appropriate.
-
         If we have a numeric type, we'll enable the range and
         tolerance widgets. Otherwise, we'll disable them.
-
         Parameters
         ----------
         gui_type_str : str
@@ -1324,7 +802,6 @@ class EqualsWidget(DesignerDisplay, EqualsMixin, DataWidget):
 class NotEqualsWidget(EqualsWidget):
     """
     Handle the NotEquals comparison.
-
     This is simply an equals widget with the not equals symbol.
     """
     def __init__(self, *args, **kwargs):
@@ -1335,10 +812,8 @@ class NotEqualsWidget(EqualsWidget):
 class GtLtBaseWidget(DesignerDisplay, DataWidget):
     """
     Base widget for comparisons like greater, less, etc.
-
     This class should be subclassed to define "symbol" and
     instantiated with the appropriate comparison data class.
-
     These comparisons have the following properties in common:
     - The only unique field is "value"
     - The comparison can be represented by a single symbol
@@ -1391,7 +866,6 @@ class LessOrEqualWidget(GtLtBaseWidget):
 class RangeWidget(DesignerDisplay, DataWidget):
     """
     Widget to handle the "Range" comparison.
-
     Contains graphical representations of what the
     range means, since it might not always be clear
     to the user what a warning range means.
@@ -1438,7 +912,6 @@ class RangeWidget(DesignerDisplay, DataWidget):
     def setup_range_widget(self) -> None:
         """
         Do all the setup required for a range widget.
-
         - Connect the text entry fields and set the dynamic expand/contract
         - Set up the inclusive checkbox
         - Set up the symbols based on the inclusive checkbox
@@ -1480,10 +953,8 @@ class RangeWidget(DesignerDisplay, DataWidget):
     def update_symbols(self, inclusive: bool) -> None:
         """
         Pick the symbol type based on range inclusiveness.
-
         Use the less than symbol if not inclusive, and the the
         less than or equals symbol if inclusive.
-
         Parameters
         ----------
         inclusive : bool
@@ -1599,19 +1070,16 @@ class RangeWidget(DesignerDisplay, DataWidget):
 class ValueRowWidget(DesignerDisplay, EqualsMixin, DataWidget):
     """
     Row widget for the "Value" dataclass used in "ValueSet".
-
     A "ValueSet" is made up of a number of "Value" objects.
     This row widget is a bit larger than the comparison or
     configuration row widgets because there will not be
     a sub-page for modifying the fields. The following
     fields are handled here:
-
     - value: PrimitiveType
     - description: str = ""
     - rtol: Optional[Number] = None
     - atol: Optional[Number] = None
     - severity: Severity = Severity.success
-
     Note that this ends up being similar to the equals widget
     due to the same rtol/atol structure.
     """
@@ -1655,13 +1123,10 @@ class ValueRowWidget(DesignerDisplay, EqualsMixin, DataWidget):
 class ValueSetWidget(DesignerDisplay, DataWidget):
     """
     Widget for modifying the unique fields in "ValueSet"
-
     The only unique field is currently "values".
-
     This is an ordered sequence of values, where the first
     value to match in the order is the result of the
     comparison.
-
     To support this ordering, this widget has a table with
     drag and drop enabled.
     """
@@ -1688,9 +1153,7 @@ class ValueSetWidget(DesignerDisplay, DataWidget):
     ) -> None:
         """
         Adds a new or existing value to the table.
-
         New values are added to the data as well.
-
         Parameters
         ----------
         checked : bool, optional
@@ -1714,7 +1177,6 @@ class ValueSetWidget(DesignerDisplay, DataWidget):
     def setup_delete_button(self, value_row: ValueRowWidget) -> None:
         """
         Set up a row's delete button.
-
         The button needs to have the correct style and functionality.
         """
         delete_icon = self.style().standardIcon(QStyle.SP_TitleBarCloseButton)
@@ -1754,7 +1216,6 @@ class ValueSetWidget(DesignerDisplay, DataWidget):
     def move_config_row(self, source: int, dest: int) -> None:
         """
         Move the row at index source to index dest.
-
         Rearanges the table and the file.
         """
         # Skip if into the same index
@@ -1776,7 +1237,6 @@ class ValueSetWidget(DesignerDisplay, DataWidget):
     def table_drop_event(self, event: QDropEvent) -> None:
         """
         Monkeypatch onto the table to allow us to drag/drop rows.
-
         Shoutouts to stackoverflow
         """
         if event.source() is self.value_table:
@@ -1793,13 +1253,10 @@ class ValueSetWidget(DesignerDisplay, DataWidget):
 class AnyValueWidget(DesignerDisplay, DataWidget):
     """
     Widget for modifying the unique fields in "AnyValue"
-
     The only unique field is currently "values".
-
     This is an unordered sequence of primitive values.
     The comparison passes if the actual value matches any
     of these primitives.
-
     This widget will have a table of values similar to the one
     used in the global values attribute in ConfigurationGroup.
     The table is used to make editing easy and to communicate
@@ -1832,9 +1289,7 @@ class AnyValueWidget(DesignerDisplay, DataWidget):
     ) -> None:
         """
         Add a new value to the table.
-
         The default value is empty string.
-
         Parameters
         ----------
         checked : bool, optional
@@ -1862,11 +1317,9 @@ class AnyValueWidget(DesignerDisplay, DataWidget):
     def on_table_edit(self, row: int, column: int) -> None:
         """
         Slot for updating the saved values when the table is edited.
-
         Regardless of which row or column is changed, we'll reconstruct
         the entire values dictionary to make sure it is serialized with
         consistent ordering.
-
         The arguments are passed by the qt signal but are unused.
         """
         if self.adding_new_row:
@@ -1927,16 +1380,12 @@ class AnyValueWidget(DesignerDisplay, DataWidget):
 class AnyComparisonWidget(DesignerDisplay, DataWidget):
     """
     Widget for modifying the unique fields in "AnyComparison"
-
     The only unique field is currently "comparisons".
-
     This is an unordered sequence of other comparisons.
     The comparison passes if any of these sub-comparisons
     passes.
-
     This widget will use the ComparisonRowWidget to fill a table,
     much like the various configuration pages.
-
     This widget will rely on the ComparisonPage to set up and
     handle the necessary sub-pages that this needs to create.
     """
@@ -1961,9 +1410,7 @@ class AnyComparisonWidget(DesignerDisplay, DataWidget):
     ) -> None:
         """
         Add a new or existing comparison to the table.
-
         New comparisons are also added to the data.
-
         Parameters
         ----------
         checked : bool, optional
@@ -1991,7 +1438,6 @@ class AnyComparisonWidget(DesignerDisplay, DataWidget):
     def setup_delete_button(self, comparison_row: ComparisonRowWidget) -> None:
         """
         Set up a row's delete button.
-
         The button needs to have the correct style and functionality.
         """
         delete_icon = self.style().standardIcon(QStyle.SP_TitleBarCloseButton)
